Bug fixes and support for 6.0.0
Most of the tests pass, but creating a core with a greater
than one replication factor seems to time out on my mac.

On the other hand, I didn't encounter this problem in a
Vagrant installation with real data, and the collection is
created with the right number of shards, etc. So there's a
note on that test for now.

Resolves #28